add UI friendly name
