+++ litellm/types/management_endpoints/router_settings_endpoints.py
@@ -14,6 +14,7 @@ class RouterSettingsField(BaseModel):
     field_description: str
     field_default: Any = None
     options: Optional[List[str]] = None  # For fields with predefined options/enum values
+    ui_field_name: str  # User-friendly display name
 
 
 # Define all available router settings fields
@@ -25,6 +26,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_description="Routing strategy to use for load balancing across deployments",
         field_default="simple-shuffle",
         options=[],  # Will be populated dynamically from Router class
+        ui_field_name="Routing Strategy",
     ),
     RouterSettingsField(
         field_name="routing_strategy_args",
@@ -32,6 +34,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Arguments to pass to the routing strategy (e.g., ttl, lowest_latency_buffer for latency-based-routing)",
         field_default={},
+        ui_field_name="Routing Strategy Args",
     ),
     RouterSettingsField(
         field_name="num_retries",
@@ -39,6 +42,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Number of retries for failed requests",
         field_default=0,
+        ui_field_name="Number of Retries",
     ),
     RouterSettingsField(
         field_name="timeout",
@@ -46,6 +50,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Timeout for requests in seconds",
         field_default=None,
+        ui_field_name="Timeout",
     ),
     RouterSettingsField(
         field_name="stream_timeout",
@@ -53,6 +58,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Timeout for streaming requests in seconds",
         field_default=None,
+        ui_field_name="Stream Timeout",
     ),
     RouterSettingsField(
         field_name="max_fallbacks",
@@ -60,6 +66,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Maximum number of fallbacks to try before exiting the call",
         field_default=5,
+        ui_field_name="Max Fallbacks",
     ),
     RouterSettingsField(
         field_name="fallbacks",
@@ -67,6 +74,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="List of fallback model mappings",
         field_default=[],
+        ui_field_name="Fallbacks",
     ),
     RouterSettingsField(
         field_name="context_window_fallbacks",
@@ -74,6 +82,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="List of fallback models for context window errors",
         field_default=[],
+        ui_field_name="Context Window Fallbacks",
     ),
     RouterSettingsField(
         field_name="content_policy_fallbacks",
@@ -81,6 +90,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="List of fallback models for content policy errors",
         field_default=[],
+        ui_field_name="Content Policy Fallbacks",
     ),
     RouterSettingsField(
         field_name="allowed_fails",
@@ -88,6 +98,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Number of times a deployment can fail before being added to cooldown",
         field_default=None,
+        ui_field_name="Allowed Fails",
     ),
     RouterSettingsField(
         field_name="cooldown_time",
@@ -95,6 +106,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Time in seconds to cooldown a deployment after failure",
         field_default=None,
+        ui_field_name="Cooldown Time",
     ),
     RouterSettingsField(
         field_name="retry_after",
@@ -102,6 +114,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Minimum time to wait before retrying a failed request in seconds",
         field_default=0,
+        ui_field_name="Retry After",
     ),
     RouterSettingsField(
         field_name="retry_policy",
@@ -109,6 +122,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Custom retry policy for different exception types",
         field_default=None,
+        ui_field_name="Retry Policy",
     ),
     RouterSettingsField(
         field_name="model_group_alias",
@@ -116,6 +130,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Aliases for model groups",
         field_default={},
+        ui_field_name="Model Group Alias",
     ),
     RouterSettingsField(
         field_name="enable_pre_call_checks",
@@ -123,6 +138,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Enable pre-call checks before routing requests",
         field_default=False,
+        ui_field_name="Enable Pre-call Checks",
     ),
     RouterSettingsField(
         field_name="default_litellm_params",
@@ -130,6 +146,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Default parameters for Router.chat.completion.create",
         field_default=None,
+        ui_field_name="Default LiteLLM Params",
     ),
     RouterSettingsField(
         field_name="set_verbose",
@@ -137,6 +154,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Enable verbose logging for router",
         field_default=False,
+        ui_field_name="Verbose Logging",
     ),
     RouterSettingsField(
         field_name="default_max_parallel_requests",
@@ -144,6 +162,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Default maximum parallel requests across all deployments",
         field_default=None,
+        ui_field_name="Max Parallel Requests",
     ),
     RouterSettingsField(
         field_name="enable_tag_filtering",
@@ -151,6 +170,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Enable filtering deployments by tags",
         field_default=False,
+        ui_field_name="Enable Tag Filtering",
     ),
     RouterSettingsField(
         field_name="disable_cooldowns",
@@ -158,6 +178,7 @@ ROUTER_SETTINGS_FIELDS: List[RouterSettingsField] = [
         field_value=None,
         field_description="Disable cooldown mechanism for failed deployments",
         field_default=None,
+        ui_field_name="Disable Cooldowns",
     ),
 ]
